docs: :memo: upgrade storybook & améliorations doc
- mise à jour vers storybook 7.5.3
- ajoute les styles du DSFR pour toutes les docs mdx
- ajoute une page pour prévisualiser toutes les couleurs du DSFR
