--- .storybook/manager.js
@@ -1,6 +1,0 @@
-import { addons } from '@storybook/addons';
-import VueDsfrTheme from './vue-dsfr-theme.js';
-
-addons.setConfig({
-  theme: VueDsfrTheme,
-});
