+++ .storybook/preview-head.html
@@ -1,4 +1,3 @@
 <script>
-  // Fix: add `jest` to global
   window.global = window
 </script>
