+++ package.json
@@ -43,12 +43,12 @@
   "scripts": {
     "check-exports": "node ./ci/check-exports.mjs",
     "check-exports-ci": "node ./ci/check-exports.mjs --ci",
-    "dev": "run-s storybook",
     "clean-dist": "rimraf dist lib",
     "prebuild": "run-s clean-dist",
     "build": "run-s build:vite typegen",
     "build:vite": "vite build",
     "demo": "vite",
+    "dev": "storybook dev -p 6006",
     "vitest": "vitest",
     "coverage": "vitest run --coverage",
     "test": "run-s test:unit test:ct-ci",
@@ -59,15 +59,13 @@
     "format": "run-s lint -- --fix",
     "lint": "eslint ./src/**/*.{vue,js,ts}",
     "lint:staged": "eslint",
-    "storybook": "cross-env NODE_OPTIONS=--openssl-legacy-provider storybook dev -p 6006",
-    "build-storybook": "NODE_OPTIONS=--openssl-legacy-provider storybook build",
-    "preview": "run-s build-storybook",
-    "view": "vite serve storybook-static",
     "prepublishOnly": "run-s build",
     "semantic-release": "semantic-release",
     "typecheck": "vue-tsc --noEmit -p tsconfig.vitest.json --composite false",
     "vite:build": "vite build && npm run typegen",
-    "typegen": "vue-tsc --declaration --emitDeclarationOnly"
+    "typegen": "vue-tsc --declaration --emitDeclarationOnly",
+    "storybook": "storybook dev -p 6006",
+    "build-storybook": "storybook build"
   },
   "dependencies": {
     "@gouvfr/dsfr": "~1.10.2",
@@ -77,45 +75,27 @@
     "vue": "^3.3.8"
   },
   "devDependencies": {
-    "@babel/core": "^7.23.3",
-    "@babel/eslint-parser": "^7.23.3",
-    "@babel/plugin-syntax-jsx": "^7.23.3",
-    "@babel/preset-env": "^7.23.3",
-    "@babel/preset-typescript": "^7.23.3",
     "@cypress/vite-dev-server": "^5.0.6",
     "@cypress/vue": "^6.0.0",
-    "@mdx-js/react": "^2.3.0",
-    "@rollup/plugin-alias": "^5.0.1",
-    "@rollup/plugin-babel": "^6.0.4",
-    "@rollup/plugin-commonjs": "^25.0.7",
-    "@rollup/plugin-node-resolve": "^15.2.3",
-    "@rollup/plugin-replace": "^5.0.5",
-    "@rollup/plugin-url": "^8.0.2",
-    "@rushstack/eslint-patch": "^1.5.1",
-    "@storybook/addon-a11y": "^7.5.2",
-    "@storybook/addon-actions": "7.5.2",
-    "@storybook/addon-docs": "7.5.2",
-    "@storybook/addon-essentials": "7.5.2",
-    "@storybook/addon-interactions": "7.5.2",
-    "@storybook/addon-links": "7.5.2",
-    "@storybook/addon-mdx-gfm": "7.5.2",
+    "@rushstack/eslint-patch": "^1.3.3",
+    "@storybook/addon-essentials": "^7.5.3",
+    "@storybook/addon-interactions": "^7.5.3",
+    "@storybook/addon-links": "^7.5.3",
     "@storybook/addon-styling": "^1.3.7",
-    "@storybook/client-api": "7.5.2",
-    "@storybook/client-logger": "7.5.2",
-    "@storybook/core-common": "7.5.2",
-    "@storybook/jest": "^0.2.3",
+    "@storybook/addon-themes": "^7.5.3",
+    "@storybook/blocks": "^7.5.3",
+    "@storybook/test": "^7.6.0-alpha.4",
     "@storybook/testing-library": "^0.2.2",
-    "@storybook/vue3": "7.5.2",
-    "@storybook/vue3-vite": "7.5.2",
-    "@testing-library/jest-dom": "^6.1.4",
-    "@testing-library/user-event": "^14.5.1",
+    "@storybook/vue3": "^7.5.3",
+    "@storybook/vue3-vite": "^7.5.3",
+    "@testing-library/jest-dom": "^6.1.2",
+    "@testing-library/user-event": "^14.4.3",
     "@testing-library/vue": "^7.0.0",
-    "@types/jest": "^29.5.8",
-    "@types/jsdom": "^21.1.5",
-    "@types/node": "^18.18.9",
-    "@typescript-eslint/eslint-plugin": "^6.10.0",
-    "@typescript-eslint/parser": "^6.10.0",
-    "@vitejs/plugin-vue": "^4.4.1",
+    "@types/jsdom": "^21.1.2",
+    "@types/node": "^18.17.14",
+    "@typescript-eslint/eslint-plugin": "^6.5.0",
+    "@typescript-eslint/parser": "^6.5.0",
+    "@vitejs/plugin-vue": "^4.3.4",
     "@vitejs/plugin-vue-jsx": "^3.0.2",
     "@vue/compiler-sfc": "^3.3.8",
     "@vue/eslint-config-standard": "^8.0.1",
@@ -128,7 +108,6 @@
     "c8": "^8.0.1",
     "chalk": "^5.3.0",
     "copyfiles": "^2.4.1",
-    "core-js": "^3.33.2",
     "cross-env": "^7.0.3",
     "cypress": "^13.5.0",
     "cypress-plugin-tab": "^1.0.5",
@@ -157,16 +136,10 @@
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "rimraf": "^5.0.5",
-    "rollup": "^3.29.4",
-    "rollup-plugin-css-only": "^4.5.2",
-    "rollup-plugin-postcss": "^4.0.2",
-    "rollup-plugin-scss": "^4.0.0",
-    "rollup-plugin-terser": "^7.0.2",
     "rollup-plugin-visualizer": "^5.9.2",
-    "rollup-plugin-vue": "^6.0.0",
-    "semantic-release": "^21.1.2",
-    "start-server-and-test": "^2.0.2",
-    "storybook": "7.5.2",
+    "semantic-release": "^21.1.1",
+    "start-server-and-test": "^2.0.0",
+    "storybook": "^7.5.3",
     "tsd-jsdoc": "^2.5.0",
     "typescript": "^5.2.2",
     "vite": "^4.5.0",
